optimize: Lazy load the gif
