+++ src/components/pages/LandingPage.js
@@ -1,9 +1,8 @@
-import React from "react"
+import React, { Suspense } from "react"
 import { SECTION_NAMES } from "../vars"
 import { NavDetailed } from "../Nav"
 
-const getImageUrl = () =>
-    Math.random() > 0.5 ? "./img/small-robot-small.gif" : "./img/small-robot-2-small.gif"
+const RandomRobotGif = React.lazy(() => import("./RandomRobotGif"))
 
 class LandingPage extends React.Component {
     pageName = SECTION_NAMES.landingPage
@@ -13,16 +12,11 @@ class LandingPage extends React.Component {
     render = () => (
         <>
             <div id="landing">
-                <img
-                    loading="lazy"
-                    src={getImageUrl()}
-                    alt="hexapod robot animation"
-                    height="75"
-                    style={{ marginTop: "20px", borderRadius: "20px" }}
-                />
+                <Suspense fallback={<p>A cute random robot coming soon...</p>}>
+                    <RandomRobotGif />
+                </Suspense>
                 <h1>Mithi's Bare Minimum Hexapod Robot Simulator</h1>
                 <p>Enjoy your stay and share with your friends!</p>
-
                 <NavDetailed />
             </div>
         </>
